KfProducer design fix
